+++ src/app/connection/executing.rs
@@ -5,6 +5,6 @@ use async_trait::async_trait;
 pub(crate) trait Executing {
     async fn execute_sqlx(
         &mut self,
-        query: String,
+        query: &str,
     ) -> Result<(Vec<String>, Vec<Vec<String>>), sqlx::Error>;
 }