+++ src/app/storage/mod.rs
@@ -1,4 +1,6 @@
 mod id;
 mod storage;
+mod editors_storage;
 
 pub(crate) use storage::*;
+pub(crate) use editors_storage::*;